add id to StockItem, in oder to have as many items with same location as i like

+++ src/main/java/com/cncstock/repository/StockItemRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface StockItemRepository extends JpaRepository<StockItem, Integer> {
+public interface StockItemRepository extends JpaRepository<StockItem, Long> {
 
     List<StockItem> findByTitleContaining(String title);
 }
